Updated Condition for Delete company

+++ src/main/java/com/tw/apistackbase/controller/CompanyController.java
@@ -3,6 +3,7 @@ package com.tw.apistackbase.controller;
 import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import com.tw.apistackbase.service.CompanyService;
+import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -17,6 +18,7 @@ import static org.springframework.http.HttpStatus.*;
 @RestController
 @RequestMapping("/companies")
 public class CompanyController {
+    public static final String NOT_FOUND = "Not Found";
     @Autowired
     CompanyService companyService;
 
@@ -36,12 +38,20 @@ public class CompanyController {
         return companyService.save(company);
     }
 
+    @ResponseStatus(OK)
     @DeleteMapping(value = "/{name}", produces = {"application/json"})
-    public ResponseEntity deleteCompanyByName(@PathVariable String name){
-        return companyService.deleteCompanyByName(name);
+    public ResponseEntity deleteCompanyByName(@PathVariable String name) throws Exception{
+        Company findCompany = companyService.findByNameContaining(name);
+        if(findCompany != null)
+            return companyService.deleteCompanyByName(name);
+        throw new NotFoundException(NOT_FOUND);
     }
     @PatchMapping(value = "/{name}",produces = {"application/json"})
-    public ResponseEntity updateCompanyInfo(@PathVariable("name") String name, @RequestBody Company company) {
-        return companyService.updateCompanyInfo(name, company);
+    public ResponseEntity updateCompanyInfo(@PathVariable("name") String name, @RequestBody Company company) throws Exception{
+        ResponseEntity updateCompany = companyService.updateCompanyInfo(name, company);
+        Company findCompany = companyService.findByNameContaining(name);
+        if(findCompany != null)
+            return updateCompany;
+        throw new NotFoundException(NOT_FOUND);
     }
 }
